Remove Mixin and Toolbar
