+++ src/index.ts
@@ -3,9 +3,4 @@ React-Quill
 https://github.com/zenoamaro/react-quill
 */
 import Component from './component';
-import Mixin from './mixin';
-import Toolbar from './toolbar';
-
 export default Component;
-export {Mixin};
-export {Toolbar};
